+++ app/src/main/java/com/example/uberapp_tim/model/users/Passenger.java
@@ -10,7 +10,7 @@ public class Passenger extends User implements Serializable {
 
     private Route favoriteRoute;
 
-    public Passenger(int id, String name, String lastName, String email, String phoneNumber,
+    public Passenger(Long id, String name, String lastName, String email, String phoneNumber,
                      String address, String password, Bitmap profilePhoto,
                      boolean blocked, Route favoriteRoute){
         super(id, name, lastName, email, phoneNumber, address, password, profilePhoto, blocked);
